Changed <parent> of model POM to nic-parent POM.
This patch modified the <parent> element of the model/pom.xml to point
to parent/pom.xml according to the best practices.
https://wiki.opendaylight.org/view/CrossProject:HouseKeeping_Best_Practices_Group:Project_layout

This patch also moved <properties> and <dependencyManagement> elements
of the root pom.xml to the parent/pom.xml for the above change.

Since the <build> element of the root pom.xml was configuration for the
model POM, this patch moved it to the model/pom.xml.

Change-Id: If02dd73b803782665cb8c73cb0c697bb229f29b4
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>