Add data change listener and intent capturing registry for further processing for GBP or VTN resolvers.
The idea here is that NicProviderDataChangeListener listens to changes to the model. Intent keeps an internal queue of intents. The order is not important. Depending on data changes, a new intent can get added to the queue, updates are requeued if they have not been processed. Deleted intents are also pulled from the queue. 

Since this Registry is a singleton, there is only one instance of this queue for each ODL running instance. I am thinking that it may work well if we have CLI code also send intent changes - additions, updates and deletes to this queue. 

The next step is to add the processors for GBP and VTN. They simply pull items from the queue and either run them right away or schedule the intent to be processed at a later if that is dictated by the constraint. The later step will taken care of by SingletonTask management module. 

All these can be abandoned if the steps described is not realizable.

Change-Id: I7bcb273a03b2d89ffec09cdc6da6071615443b6a
Signed-off-by: Ben Eze <beze@inocybe.com>