Update patch set 9
Patch Set 9:

(1 comment)

If I understand this patch set correctly you are creating an internal queue for processing intents. Thus the data change handler is managing the queue.

The problem I see with this approach is that this means the NIC feature now has to manage it own thread pool, which is a thread pool that is separate from those that the MD-SAL / controller use to manage notifications, data changes etc. 

To be honest, I like the approach, what I don't like is that the MD-SAL is effectively forcing all features to manage their own threads as opposed to have a general way to manage processing and allot processing to various capabilities.

I like this approach because it effectively ensures that the MD-SAL data change processing threads are "blocked" while an intent is processed. 

I would recommend that if NIC has to manage its own thread pool, that the size of that pool be controlled by a configuration parameter.

Patch-set: 9
Reviewer: Gerrit User 1045 <1045@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0