Update patch set 11
Patch Set 11: Code-Review+1

I made the change set 11.
I've confirmed that this change does not break the VTN Renderer.

There are three changes in the change set 11 from the change set 10.
(1) Updated commit message according to the following guide http://git-scm.com/book/ch5-2.html#Commit-Guidelines. The guideline says that the first line shoud be 50 chars or less, and the blank line is needed after the first line.
(2) Using "git mv" to move the 89-vtn-renderer.xml. It make it clear that this patch just rename the file.
(3) Changed the final name of the config file to etc/opendaylight/karaf/89-vtn-renderer.xml.

Patch-set: 11