[WIP][DO NOT MERGE] Add and fix config-pusher for intentengine-impl.
This patch add the config-pusher to the feature file under odl-nic-core feature.
When testing this modification, I had config pusher exception. So the following try
to fix the exception:
- add the intentengine-impl config file dependency to the feature pom file
- rename the config file to default-config.xml
- rename the namespace to urn/opendaylight/nic instead of urn/opendaylight/params/xml/ns/yang
- rename the intentengine-impl-config.yang to intentengine-impl.yang
- modidy the yang file to reflect intentengine-impl instead of intentengine-impl-config
- remove useless folder/config file: src/main/resources/config/default-config.xml

This is the exception I've been struggling for: https://gist.github.com/adetalhouet/95168c6e11958de171ce
Indeed, not very helpful.
The config file is well pushed in distribution/target/assembly/etc/opendaylight/karaf.
IMHO there is no typo.

I won't be there on Monday so if you want it done by then, feel free to hack on it and hijack my patch.
But this can get in as a fix after the code freeze.

Patch set 1: commit the fix
Patch set 2: rebase on master
Patch set 3: update the comment

Change-Id: I4062edf9e0d1319a275bf241909b80be157ae73a
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>