Fix converting error of VTN Renderer.
First, VTN Renderer always output an error message immediately after its
initialization. The root cause of this issue was that VTN Renderer did
not do null check, and attempted to use null.
Second, VTN Renderer failed to convert an intent to VTN configuration in
the case that an intent was updated, since the way of handling the type
of data change event was wrong.

This patch fix these issues, and also removes unnecessary imports.

Change-Id: I2c1c0d9f1a768271363c6cbb0c06a9e04ddd2d2b
Signed-off-by: Hideyuki Tai <Hideyuki.Tai@necam.com>