Update patch set 2
Patch Set 2:

This approach hides all the data transaction, etc exceptions sourced from the MDSAL from the higher level functions. Is this really what you want to do?

This means the high level logic has no chance to react to the exception and decide what the correct action might be. You might end up with hidden errors in the runtime and a controller left in an error state because the exception is squelched.

Patch-set: 2