Fix bundle compilation due to unresolved checkstyle-licence.txt
Within NIC project, each of the bundle defines the checkstyle policy. (This might not be a good way to do it, but still it's like that).
Thus when comnpiling one bundle only, compilation is lookling for checkstyle-licence.txt in the bundle that is beeing compiled.
But the file is at the root of the project, so the compilation failed complaining about missing file.
This patch fix this issue by removing the checkstyle-licence.txt file and by using the global variable ${checkstyle.header.file} instead.

Change-Id: I54d4daef88c307722b90b7e2f6c2edc03be4ad7d
Signed-off-by: Alexis de Talhouët <adetalhouet@inocybe.com>