Update patch set 1
Patch Set 1: Code-Review-1

So you get the license checkstyle for free now from odlparent. So you don't actually need to define a license checker anymore as odlparent is not providing it globally.

I think a better update to this patch is to remove the extra license checker altogether.

Patch-set: 1
Label: Code-Review=-1