Update patch set 2
Patch Set 2:

I believe I was missing 2 things:
- the 'default' execution ID that I was overwriting
- I was expected the build to fail on violation (but this is not define like that in odlparent).

So I guess I can just create a new execution ID to avoid overwriting the one from odlparent and create a new one to add the "fail on violation" rule. 

Thank you for the taking the time to explain the usage to me.

Patch-set: 2