Update patch set 4
Patch Set 4: Code-Review-1

I took a look, but I agree with Gabriel. The trick part is that the intent is very close to the end-user and far away from the implementation and for end-users the use of MPLS is not visible directly.
You can assume it's been used when doing a traceroute and when you see that a packet hops between two very distant nodes. I think there's a decision that is taken somewhere inside the intent, probably the renderer.

We had a discussion on the irc channel about it. So you know more details.

Patch-set: 4
Label: Code-Review=-1