Update patch set 1
Patch Set 1:

Hello Thanh, I know it's a bit late to merge in Beryllium branch, although I'm wondering if you could press the submit button for this one if not too late. 
It was not correctly advertised on our end, thus us being late, but this patch fixes NIC's VTN CSIT job.

Patch-set: 1