Update patch set 9
Patch Set 9: Code-Review-1

(14 comments)

Please use curly brackets for function/for/if/while/... Do not use one line for a if, or a function or whatever. Always expand the code and add curly brackets. That is ODL guideline.

A lot of functions added in IntentAddShellCommand could be moved to some utility class, maybe CLIUtils or whatever.
Also, provide UT for those methods.
Finally, when throwing exception, make sure to add the thrown exception in the message, so we can easily debug.
Thanks

Patch-set: 9
Label: Code-Review=-1