Update patch set 2
Patch Set 2:

I'm having hard time getting the problem. SFT is failing because of:

13:28:24 java.io.IOException: Error resolving artifact org.opendaylight.sfc:features-sfc:xml:features:0.4.0-SNAPSHOT: Could not find artifact org.opendaylight.sfc:features-sfc:xml:features:0.4.0-SNAPSHOT

Which confuse me because this patch is moving sfc version to 0.5.0-SNAPShOT, hence it shouldn't look for 0.4.0-SNAPSHOT. And I couldn't find any other occurrence of the sfc version, beside the two modified in this patch.

Patch-set: 2